Make DataReader implements AutoClosable
Make DataReader implements AutoClosable. Implement close and NetcdfDataReader. Call done() on DSS accessors. Make TemporalDataReader implements AutoClosable. Update NetcdfDataReader resources to close properly. Update NetcdfDataWriterTest and TemporalDataReaderTest to close resources.
