+++ vortex-api/src/main/java/mil/army/usace/hec/vortex/io/AscDataReader.java
@@ -310,4 +310,9 @@ class AscDataReader extends DataReader {
                 .map(VortexDataInterval::of)
                 .toList();
     }
+
+    @Override
+    public void close() throws Exception {
+        // No op
+    }
 }
